Moved the kafka metadata fetcher in a dedicated class. Renamed service to metadata_kafka, updated README and configuration properties accrodingly.
