+++ src/run/conf/configuration.properties
@@ -21,4 +21,4 @@ consumer.group.id=ztesting
 consumer.message.unit.print=true
 
 ## metadata fetchin settings
-metadata.describe.cluster=true
+metadata_kafka.describe.cluster=true
